TestXMLIO: Make integers() take count as long, not int.
To match the parameter of the method it is passed to.

+++ model/src/test/java/legacy/xmlio/TestXMLIO.java
@@ -324,7 +324,7 @@ public final class TestXMLIO {
 	}
 
 	// TODO: Reformat methods using these extracted helpers
-	private static Stream<Integer> integers(final int count) {
+	private static Stream<Integer> integers(final long count) {
 		return SINGLETON_RANDOM.ints(count).boxed();
 	}
